requirements.txt
pandas
matplotlib
groq
sentence-transformers
numpy
python-dotenv
